Fix deallocation size

+++ tests/disk/disk_compute.c
@@ -165,8 +165,8 @@ int dotest(struct starpu_disk_ops *ops, char *base)
 			try = 0;
 		}
 
-	starpu_free_flags(A, NX*sizeof(double), STARPU_MALLOC_COUNT);
-	starpu_free_flags(C, NX*sizeof(double), STARPU_MALLOC_COUNT);
+	starpu_free_flags(A, NX*sizeof(int), STARPU_MALLOC_COUNT);
+	starpu_free_flags(C, NX*sizeof(int), STARPU_MALLOC_COUNT);
 
 	unlink(path_file_start);
 	unlink(path_file_end);
@@ -186,8 +186,8 @@ int dotest(struct starpu_disk_ops *ops, char *base)
 enodev:
 	return STARPU_TEST_SKIPPED;
 enoent2:
-	starpu_free_flags(A, NX*sizeof(double), STARPU_MALLOC_COUNT);
-	starpu_free_flags(C, NX*sizeof(double), STARPU_MALLOC_COUNT);
+	starpu_free_flags(A, NX*sizeof(int), STARPU_MALLOC_COUNT);
+	starpu_free_flags(C, NX*sizeof(int), STARPU_MALLOC_COUNT);
 enoent:
 	FPRINTF(stderr, "Couldn't write data: ENOENT\n");
 	starpu_shutdown();
